refactor(core): split ColumnTable modules and long pipeline test suites
Extract table-types and table-coerce from table.ts so free inference/
coercion helpers are separate from the ColumnTable class; re-export the
public surface from table.ts for a stable import path.

Replace the facets and pipeline-position-transform characterization
monoliths with focused suites under tests/facets/ and
tests/pipeline-position-transform/, sharing wrapSpec/log10Rows fixtures.
